[fenix] For https://github.com/mozilla-mobile/fenix/issues/12797: Use measured height/width for loading a thumbnail
Using `onNextLayout` has the side-effect of showing the thumbnail with
the tab.id from a recycled view.

The root cause of needing that call is to retrieve the measured
height/width from the view first before requesting. A simpler solution
to avoid the complexity is to retrieve these values from the dimen
resources.

Co-authored-by: Gabriel Luong <gabriel.luong@gmail.com>