[fenix] For https://github.com/mozilla-mobile/fenix/issues/15644: change apply to with in GleanMetricsService.
While the callback receiver is identical in these two methods, they're
semantically different: apply is for initializing the receiver while with
is anything else benefiting from a new `this` receiver.

I didn't change the usage of apply that has a return statement because I
was afraid my change might change behavior.