[fenix] For https://github.com/mozilla-mobile/fenix/issues/23500: creditcards: focus cardNumberInput in createView
Right now, when the user navigates to the "Add Card" screen, the "Card Number"
title will be focused by default. Instead, let's make it so that the first
field for data entry is focused.

Since we're focusing the first field, let's also make sure that the soft
keyboard opens and that the cursor is placed at the end of whatever text has
already been input.