Bug 1807531 - Honor private_browsing_mode value from external intents Change IntentReceiverActivity so that if the incoming intent has the private_browsing_mode extra set to true, it overrides the openLinksInAPrivateTab setting. This allows external apps to set this flag for sites they want opened in a private app explicitly.
External apps cannot force the link to open in a non-private tab if openLinksInAPrivateTab is enabled, in that case the extra will be ignored.

This is similar to an old fennec feature ( https://bugzilla.mozilla.org/show_bug.cgi?id=1347583 ), however I used the "private_browsing_mode" extra name since it already appears in the code, instead of the "private_tab" extra that fennec used.

Bugzilla bug is 1807531, GitHub bug is #26158