Bug 1812149 - Add ActivityContextWrapper before updating locale
From ohall's learnings, we know that we need to use the original Android
Activity Context with the system PrintManager. When a configuration
change happens, a new Context is created and the original is lost.
The most straight-forward way to do that is to use a ContextWrapper
which internally holds a reference to the original context.