Bug 1861173 - Change timing of the Review Checker onboarding CFR
The current onboarding CFR is set to show the first time someone
lands on a product detail page. If the user dismisses the CFR and
does not opt-in, we show it on the next product detail page after at
least 24 hours have elapsed.
This patch changes this to showing the CFR up to 3X (instead of 2)
with at least 12 hrs in-between (instead of 24).