Bug 1868886: Fix Nightly Font Telemetry
Unfortunately, the data we got back was flawed; and I hadn't noticed:
the device information was present, but the font data was just populating
the errors array instead of the fonts array. The FileInputStream on
live devices doesn't support reset, so it's been replaced with two separate
FileInputStreams.

Also bump the integer to cause another resubmission from Nightly users,
and include the submission integer.