Bug 1888766 - Decouple the dismiss and profiler state update logics r=mleclair,android-reviewers,gl
Previously we had the `profilerViewModel.setProfilerState` call inside the
dismiss method. Because of this it was updating the profiler state prematurely
when user was pressing "upload profile" button and then returning home quickly.

After decoupling the dismiss and setProfilerState calls, now we only update the
profiler state once the profiler state changes, which makes sure that profiler
states in geckoview and ProfilerViewModel stay in sync.

Differential Revision: https://phabricator.services.mozilla.com/D206647