Bug 1807268 - Fix verifyOpenAllInPrivateTabsTest UI test r=ohorvath
The UI test was disabled a while ago due to the Home Activity ANR.
This always happened after performing the close tabs action.

After taking a closer look, I've noticed that we tried to open the main menu robot instead of the home screen robot (closing all the tabs will result in a transition to the home screen)
I've also noticed that the closing of the open tabs action was redundant, and wasn't really needed to test this specific scenario so I've removed it.

✅ The UI test successfully passed 200x on Firebase.

Differential Revision: https://phabricator.services.mozilla.com/D208505