Bug 1879011 - Fix sponsored suggest related UI tests r=ohorvath
Both **openFirefoxSuggestSponsoredSearchResultsTest** and **verifyFirefoxSuggestSponsoredSearchResultsWithEditedKeywordTest** were flaky or even failed because no suggestion was displayed in the search wrapper.

This problems most likely was caused by [[ https://bugzilla.mozilla.org/show_bug.cgi?id=1813587 | 1813587 ]].

To overcome the problems, I've added 2 extra steps to all existing sponsored suggestions related UI tests, in which we use DuckDuckGo which proved to work properly to assert and interact with the suggestions.

✅ All 4 UI tests successfully passed 150x on Firebase:

  - **verifyFirefoxSuggestSponsoredSearchResultsTest**
  - **verifyFirefoxSuggestSponsoredSearchResultsWithPartialKeywordTest**
  - **openFirefoxSuggestSponsoredSearchResultsTest**
  - **verifyFirefoxSuggestSponsoredSearchResultsWithEditedKeywordTest**

Differential Revision: https://phabricator.services.mozilla.com/D211346