Bug 1848623 - Fix verifyErrorMessagesForInvalidSearchEngineUrlsTest UI test r=aaronmt
The UI test was recently flaky when trying to verify the custom search engine error message.

Switched from using Espresso to UIAutomator (offers the possibility to wait for an element, in our case the response that should display the error message) and the UI test successfully passed 200x on Firebase. ✅

Differential Revision: https://phabricator.services.mozilla.com/D214984