Bug 1901795 - New jump back in CFR related UI test r=ohorvath
New UI test that checks that while the search fragment is enabled the jump back in CFR is not displayed.
There was also a crash encountered in the past related this matter and hopefully the UI test will catch any regression in the future.

The UI test successfully passed 100x on Firebase ✅

Differential Revision: https://phabricator.services.mozilla.com/D215349