Bug 1902798 - part 1: Enable navbar in nightly r=android-reviewers,skhan
Changing nimbus defaults to enable the feature in nighly. The UI change failed many UI tests, disabled the feature for UI tests for now. The verifyRunBlockingAndStrictModeSuppresionCount test ignores all the flags and it reported a regression: Menu button was being created twice, for both Navbar and Toolbar. Refactored HomeFragment to not cause the regression.

Differential Revision: https://phabricator.services.mozilla.com/D214428