Bug 1911085 - Avoid anchoring the snackbar to top toolbars r=android-reviewers,rsainani, a=dmeehan
Recently we introduced a new dependency for the snackbars - `toolbarLayout`
Which on tablets can appear at the top, similarly to the old `toolbar` dependency.

Previously this wasn't found and the snackbar would default to be placed at the bottom
of the screen, now we specifically place it as such when using any top toolbar.

Differential Revision: https://phabricator.services.mozilla.com/D218326