Bug 1908351 - Fix keyboard popup after changing app theme r=android-reviewers,tchoh
Since we navigate through NavContorller to all the fragments in our repository, any bundle comes from navArgs and not necessarily savedInstanceState. However activity.recreate() called after changing app theme populates the navArgs with the cached data. That is if HomeFragment was open with FOCUS_ON_ADDRESS_BAR as true from some fragment, then navigated to settings and changed the app theme, FOCUS_ON_ADDRESS_BAR becomes true again. However this also makes the savedInstanceState as not null, essentially an empty Bundle Object. This is how we know, that activity.recreate() was called and therefore we don't want the keyboard to popup, hence marking it as false.

Differential Revision: https://phabricator.services.mozilla.com/D219585