Bug 1920702 - Correctly record is_first_run as boolean r=chumphreys,android-reviewers,jonalmeida
Bug 1898552 added the recorded context, which recorded the "isFirstRun" param
as a boolean. However, the recorded context is splatted into the targeting
attributes inside the Nimbus SDK [1] and overrides other targeting members,
which include `isFirstRun: String` and `is_first_run: bool`. This has resulted
in targeting evaluation that checks against `isFirstRun == "'true'"` failing,
causing first run experiments to not enroll correctly. Recording the value as
`is_first_run` instead, which is expected to be a boolean, addresses the issue.


[1]: https://github.com/mozilla/application-services/blob/b10c1319e74041b2777e07657fc83dc879f841f4/components/nimbus/src/stateful/evaluator.rs#L22-L23

Differential Revision: https://phabricator.services.mozilla.com/D223399